// src/types/file/Extension.ts
export enum Extension {
	// Image extensions
	GIF = '.gif',
	JPEG = '.jpeg',
	JPG = '.jpg',
	PNG = '.png',
	SVG = '.svg',

	// Microsoft Word documents
	DOC = '.doc',
	DOCX = '.docx',

	// Text documents
	TXT = '.txt',
	PDF = '.pdf',
	CSV = '.csv',

	// Microsoft Excel and similar
	XLS = '.xls',
	XLSX = '.xlsx',

	// Microsoft PowerPoint
	PPT = '.ppt',
	PPTX = '.pptx',
}