// src/types/file/FileType.ts
export enum FileType {
	// Image extensions
	GIF = 'image/gif',
	JPEG = 'image/jpeg',
	JPG = 'image/jpg',
	PNG = 'image/png',
	SVG = 'image/svg+xml',

	// Microsoft Word documents
	DOC = 'application/msword',
	DOCX = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',

	// Text documents
	TXT = 'text/plain',
	PDF = 'application/pdf',
	CSV = 'text/csv',

	// Microsoft Excel and similar
	XLS = 'application/vnd.ms-excel',
	XLSX = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',

	// Microsoft PowerPoint
	PPT = 'application/vnd.ms-powerpoint',
	PPTX = 'application/vnd.openxmlformats-officedocument.presentationml.presentation',
}